Re-introduce leave process loop
By forcing the process to leave, the potential for a race condition is eliminated where an asynchronous process could get inconsistent results from Get Lifecycle State before the current state has been handled completely in the main process loop. This is an unlikely scenario, but would be tedious to debug.